Add files via upload
This is the program I created from the final assignment in CS344. It uses daemon programs to encode and decode plaintext files using keys that can be generated at will. This is also known as OTP, One time pad. Its best if you just use the p4gradingscript to run it. (grading script is BASHscript and the program are a collection of C files).